Move screen reader message hook, focus helpers and errorToInputInfo to shared libs
