+++ frontend/src/citizen-frontend/applications/editor/contact-info/ChildSubSection.tsx
@@ -6,6 +6,7 @@ import React from 'react'
 
 import Checkbox from 'lib-components/atoms/form/Checkbox'
 import InputField from 'lib-components/atoms/form/InputField'
+import { errorToInputInfo } from 'lib-components/input-info-helper'
 import AdaptiveFlex from 'lib-components/layout/AdaptiveFlex'
 import {
   FixedSpaceColumn,
@@ -17,7 +18,6 @@ import DatePicker from 'lib-components/molecules/date-picker/DatePicker'
 import { H3, Label } from 'lib-components/typography'
 import { Gap } from 'lib-components/white-space'
 
-import { errorToInputInfo } from '../../../input-info-helper'
 import { useLang, useTranslation } from '../../../localization'
 
 import type { ContactInfoSectionProps } from './ContactInfoSection'
