+++ frontend/src/citizen-frontend/applications/editor/service-need/ServiceTimeSubSectionPreschool.tsx
@@ -16,6 +16,7 @@ import { useUniqueId } from 'lib-common/utils/useUniqueId'
 import Checkbox from 'lib-components/atoms/form/Checkbox'
 import Radio from 'lib-components/atoms/form/Radio'
 import TimeInput from 'lib-components/atoms/form/TimeInput'
+import { errorToInputInfo } from 'lib-components/input-info-helper'
 import {
   FixedSpaceColumn,
   FixedSpaceRow
@@ -33,7 +34,6 @@ import {
   applicationAttachment
 } from '../../../attachments/attachments'
 import { deleteAttachmentMutation } from '../../../attachments/queries'
-import { errorToInputInfo } from '../../../input-info-helper'
 import { useLang, useTranslation } from '../../../localization'
 
 import type { ServiceNeedSectionProps } from './ServiceNeedSection'
