+++ frontend/src/citizen-frontend/calendar/CalendarPage.tsx
@@ -11,6 +11,7 @@ import FiniteDateRange from 'lib-common/finite-date-range'
 import type { CitizenCalendarEvent } from 'lib-common/generated/api-types/calendarevent'
 import LocalDate from 'lib-common/local-date'
 import { constantQuery, useQueryResult } from 'lib-common/query'
+import { focusElementOnNextFrame } from 'lib-common/utils/focus'
 import Main from 'lib-components/atoms/Main'
 import { ContentArea } from 'lib-components/layout/Container'
 import { Desktop, RenderOnlyOn } from 'lib-components/layout/responsive-layout'
@@ -25,7 +26,6 @@ import { renderResult, UnwrapResult } from '../async-rendering'
 import { useUser } from '../auth/state'
 import { useTranslation } from '../localization'
 import useTitle from '../useTitle'
-import { focusElementOnNextFrame } from '../utils/focus'
 
 import AbsenceModal from './AbsenceModal'
 import ActionPickerModal from './ActionPickerModal'
