+++ frontend/src/citizen-frontend/calendar/reservation-modal/TimeInputs.tsx
@@ -12,6 +12,7 @@ import {
   useFormField,
   useFormUnion
 } from 'lib-common/form/hooks'
+import { focusElementOnNextFrame } from 'lib-common/utils/focus'
 import { ScreenReaderOnly } from 'lib-components/atoms/ScreenReaderOnly'
 import {
   IconOnlyButton,
@@ -23,12 +24,11 @@ import {
   ExpandingInfoBox,
   InfoButton
 } from 'lib-components/molecules/ExpandingInfo'
+import { useScreenReaderMessage } from 'lib-components/utils/useScreenReaderMessage'
 import { faPlus, fasUserMinus, faTrash, faUserMinus } from 'lib-icons'
 
 import { useTranslation } from '../../localization'
-import { focusElementOnNextFrame } from '../../utils/focus'
 import TimeRangeInput from '../TimeRangeInput'
-import { useScreenReaderMessage } from '../hooks'
 
 import type { LimitedLocalTimeRangeField, ReadOnlyState } from './form'
 import {
