+++ frontend/src/citizen-frontend/decisions/decisions-page/Decisions.tsx
@@ -15,6 +15,7 @@ import type { ApplicationId } from 'lib-common/generated/api-types/shared'
 import type LocalDate from 'lib-common/local-date'
 import { formatPersonName } from 'lib-common/names'
 import { useQueryResult } from 'lib-common/query'
+import { focusElementAfterDelay } from 'lib-common/utils/focus'
 import { scrollToElement } from 'lib-common/utils/scrolling'
 import HorizontalLine from 'lib-components/atoms/HorizontalLine'
 import IconChip from 'lib-components/atoms/IconChip'
@@ -32,7 +33,6 @@ import { renderResult } from '../../async-rendering'
 import { childrenQuery } from '../../children/queries'
 import { useTranslation } from '../../localization'
 import useTitle from '../../useTitle'
-import { focusElementAfterDelay } from '../../utils/focus'
 import { decisionsQuery, financeDecisionsQuery } from '../queries'
 import { iconPropsByStatus } from '../shared'
