+++ frontend/src/employee-frontend/components/child-information/CreateApplicationModal.tsx
@@ -19,6 +19,7 @@ import { useMutationResult } from 'lib-common/query'
 import Select from 'lib-components/atoms/dropdowns/Select'
 import Checkbox from 'lib-components/atoms/form/Checkbox'
 import Radio from 'lib-components/atoms/form/Radio'
+import { errorToInputInfo } from 'lib-components/input-info-helper'
 import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
 import { PersonName } from 'lib-components/molecules/PersonNames'
 import DatePicker from 'lib-components/molecules/date-picker/DatePicker'
@@ -30,7 +31,6 @@ import { faFileAlt } from 'lib-icons'
 import { getEmployeeUrlPrefix } from '../../constants'
 import { useTranslation } from '../../state/i18n'
 import { UIContext } from '../../state/ui'
-import { errorToInputInfo } from '../../utils/validation/input-info-helper'
 import CreatePersonInput from '../common/CreatePersonInput'
 import {
   DbPersonSearch as PersonSearch,
