+++ frontend/src/employee-frontend/components/unit/tab-groups/groups/GroupModal.tsx
@@ -9,6 +9,7 @@ import type { DaycareId } from 'lib-common/generated/api-types/shared'
 import type LocalDate from 'lib-common/local-date'
 import { cancelMutation } from 'lib-common/query'
 import InputField from 'lib-components/atoms/form/InputField'
+import { errorToInputInfo } from 'lib-components/input-info-helper'
 import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
 import { MessageBox } from 'lib-components/molecules/MessageBoxes'
 import DatePicker from 'lib-components/molecules/date-picker/DatePicker'
@@ -21,7 +22,6 @@ import { fasExclamation } from 'lib-icons'
 
 import { useTranslation } from '../../../../state/i18n'
 import { UIContext } from '../../../../state/ui'
-import { errorToInputInfo } from '../../../../utils/validation/input-info-helper'
 import { allPropertiesTrue } from '../../../../utils/validation/validations'
 import { createGroupMutation } from '../../queries'
